+++ src/main/java/com/techwave/olol/auth/RestAuthenticationEntryPoint.java
@@ -1,0 +1,26 @@
+package com.techwave.olol.auth;
+
+import java.io.IOException;
+
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+
+import com.techwave.olol.global.exception.ApiException;
+import com.techwave.olol.global.exception.Error;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class RestAuthenticationEntryPoint implements AuthenticationEntryPoint {
+
+	@Override
+	public void commence(HttpServletRequest request, HttpServletResponse response,
+		AuthenticationException authException) throws IOException, ServletException {
+		log.error("[RestAuthenticationEntryPoint] : {}", request);
+
+		throw new ApiException(Error.UNAUTHORIZED);
+	}
+}
