+++ src/main/java/com/techwave/olol/auth/config/KakaoProperties.java
@@ -1,0 +1,17 @@
+package com.techwave.olol.auth.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Configuration
+@ConfigurationProperties(prefix = "kakao")
+public class KakaoProperties {
+	private String clientId;
+	private String clientSecret;
+	private String redirectUri;
+}
