+++ src/main/java/com/techwave/olol/auth/constant/GrantType.java
@@ -1,0 +1,15 @@
+package com.techwave.olol.auth.constant;
+
+import lombok.Getter;
+
+@Getter
+public enum GrantType {
+	CUSTOM("custom"),
+	KAKAO("kakao");
+
+	private String name;
+
+	GrantType(String name) {
+		this.name = name;
+	}
+}
