+++ src/main/java/com/techwave/olol/auth/dto/TokenDto.java
@@ -1,0 +1,23 @@
+package com.techwave.olol.auth.dto;
+
+import java.sql.Timestamp;
+import java.time.LocalDateTime;
+import java.util.Date;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+public class TokenDto {
+
+	private String token;
+
+	private LocalDateTime expiration;
+
+	@Builder
+	public TokenDto(String token, Date expiration) {
+		this.token = token;
+		this.expiration = new Timestamp(expiration.getTime()).toLocalDateTime();
+		;
+	}
+}
