+++ src/main/java/com/techwave/olol/auth/dto/reponse/KakaoUserInfoResponse.java
@@ -1,0 +1,16 @@
+package com.techwave.olol.auth.dto.reponse;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.Data;
+
+@Data
+public class KakaoUserInfoResponse {
+
+	private Long id;
+
+	@JsonProperty("kakao_account")
+	private Object kakaoAccout;
+
+	private Object properties;
+}
