--- src/main/java/com/techwave/olol/login/auth/TokenAccessDeniedHandler.java
@@ -1,28 +1,0 @@
-package com.techwave.olol.login.auth;
-
-import java.io.IOException;
-
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import com.techwave.olol.global.exception.ApiException;
-import com.techwave.olol.global.exception.Error;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@Component
-public class TokenAccessDeniedHandler implements AccessDeniedHandler {
-
-	@Override
-	public void handle(HttpServletRequest request, HttpServletResponse response,
-		AccessDeniedException accessDeniedException) throws IOException, ServletException {
-		log.error("[TokenAccessDeniedHandler] : {}", request);
-
-		throw new ApiException(Error.ACCESS_DENIED);
-	}
-}
