--- src/main/java/com/techwave/olol/login/config/KakaoProperties.java
@@ -1,17 +1,0 @@
-package com.techwave.olol.login.config;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Configuration
-@ConfigurationProperties(prefix = "kakao")
-public class KakaoProperties {
-	private String clientId;
-	private String clientSecret;
-	private String redirectUri;
-}
