--- src/main/java/com/techwave/olol/login/constant/AuthType.java
@@ -1,16 +1,0 @@
-package com.techwave.olol.login.constant;
-
-import lombok.Getter;
-
-@Getter
-public enum AuthType {
-
-	CUSTOM("custom"),
-	KAKAO("kakao");
-
-	private String name;
-
-	AuthType(String name) {
-		this.name = name;
-	}
-}
