--- src/main/java/com/techwave/olol/login/controller/TestController.java
@@ -1,15 +1,0 @@
-package com.techwave.olol.login.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@RequestMapping("/test")
-@Controller
-public class TestController {
-
-	@GetMapping("/kakao")
-	public String kakao() {
-		return "index";
-	}
-}
