--- src/main/java/com/techwave/olol/login/dto/TokenDto.java
@@ -1,23 +1,0 @@
-package com.techwave.olol.login.dto;
-
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-import java.util.Date;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-public class TokenDto {
-
-	private String token;
-
-	private LocalDateTime expiration;
-
-	@Builder
-	public TokenDto(String token, Date expiration) {
-		this.token = token;
-		this.expiration = new Timestamp(expiration.getTime()).toLocalDateTime();
-		;
-	}
-}
