--- src/main/java/com/techwave/olol/login/repository/RefreshTokenRepository.java
@@ -1,12 +1,0 @@
-package com.techwave.olol.login.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.techwave.olol.login.domain.RefreshToken;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, String> {
-
-	Optional<RefreshToken> findByUserId(String userId);
-}
